Revert "Update to kernel 4.19.295"
This reverts commit a3e24e5bef5e6a317f002ef791858386caab738f.

+++ fs/omfs/bitmap.c
@@ -1,4 +1,3 @@
-// SPDX-License-Identifier: GPL-2.0
 #include <linux/kernel.h>
 #include <linux/fs.h>
 #include <linux/buffer_head.h>
@@ -160,7 +159,7 @@ int omfs_allocate_range(struct super_block *sb,
 	goto out;
 
 found:
-	*return_block = (u64) i * bits_per_entry + bit;
+	*return_block = i * bits_per_entry + bit;
 	*return_size = run;
 	ret = set_run(sb, i, bits_per_entry, bit, run, 1);
